+++ rollup.test.config.js
@@ -2,18 +2,11 @@ import svelte from 'rollup-plugin-svelte';
 import resolve from '@rollup/plugin-node-resolve';
 import typescript from "@rollup/plugin-typescript";
 import autoPreprocess from 'svelte-preprocess';
-import pkg from './package.json';
-
-const name = pkg.name
-    .replace(/^(@\S+\/)?(svelte-)?(\S+)/, '$3')
-    .replace(/^\w/, m => m.toUpperCase())
-    .replace(/-\w/g, m => m[1].toUpperCase());
 
 export default {
-    input: 'src/index.ts',
+    input: 'tests/src/App.svelte',
     output: [
-        { file: pkg.module, 'format': 'es' },
-        { file: pkg.main, 'format': 'umd', name }
+        { file: 'tests/build/app.js', 'format': 'iife', name: 'app' }
     ],
     plugins: [
         svelte({