+++ src/index.ts
@@ -15,22 +15,56 @@
  * OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-import { writable } from "svelte/store"
+import {Writable, writable} from "svelte/store"
+
+/**
+ * Undoable store.
+ *
+ * A writable store with history.
+ */
+export interface UndoableStore<T> extends Writable<T> {
+    /**
+     * Change the value to the previous saved state
+     */
+    undo(): void,
+
+    /**
+     * Change the value to the next saved state
+     */
+    redo(): void,
+
+    /**
+     * Indicate if the store value can be revert to a previous state
+     */
+    canUndo(): boolean,
+
+    /**
+     * Indicate if the store value can be change to a next state
+     */
+    canRedo(): boolean,
+    /**
+     * Revert the value of the store to the oldest state.
+     * If the parameter is `true`, then the store state history is cleared
+     * @param {boolean?} clear If `true` the history is cleared
+     */
+    reset(clear?: boolean): void,
+    length(): number
+}
 
 /**
  * Create a store with undo/redo feature
  * @param {*} initial The initial value of the store
- * @param {number} [capacity=0] The maximum number of entry to remember (any number lower than 2 is considered as infinite size)
- * @param {function(*: newValue): boolean} [accept] The validation function to accept a value to be save in memory.<br/>
+ * @param {number?} capacity The maximum number of entry to remember (any number lower than 2 is considered as infinite size)
+ * @param {function(*: newValue): boolean} accept The validation function to accept a value to be save in memory.<br/>
  *                                                  Take the new store value as parameter, should return a boolean (`true` to save the value, `false` to dismiss it).<br/>
  *                                                  If ignore, it will accept all value
- * @return {Writable<*>}
+ * @return {UndoableStore<*>}
  */
-export const undoable = (initial, capacity, accept) => {
+export function undoable<T>(initial: T, capacity?: number, accept?: (newValue: T) => boolean): UndoableStore<T> {
     let states = [initial]
     let inAction = true
     let index = 0
-    let store = writable(initial)
+    const store: Writable<T> = writable(initial)
     if (typeof accept !== "function") {
         accept = () => true
     }
@@ -56,7 +90,7 @@ export const undoable = (initial, capacity, accept) => {
      * @internal
      * Change the value to the previous saved state
      */
-    store.undo = () => {
+    const undo = () => {
         inAction = true
         if (index > 0) {
             index--
@@ -69,7 +103,7 @@ export const undoable = (initial, capacity, accept) => {
      * @internal
      * Change the value to the next saved state
      */
-    store.redo = () => {
+    const redo = () => {
         inAction = true
         if (index < states.length - 1) {
             index++
@@ -83,7 +117,7 @@ export const undoable = (initial, capacity, accept) => {
      * Indicate if the store value can be revert to a previous state
      * @return {boolean}
      */
-    store.canUndo = () => {
+    const canUndo = () => {
         return index > 0
     }
 
@@ -92,7 +126,7 @@ export const undoable = (initial, capacity, accept) => {
      * Indicate if the store value can be change to a next state
      * @return {boolean}
      */
-    store.canRedo = () => {
+    const canRedo = () => {
         return index < states.length - 1
     }
 
@@ -100,9 +134,9 @@ export const undoable = (initial, capacity, accept) => {
      * @internal
      * Revert the value of the store to the oldest state.
      * If the parameter is `true`, then the store state history is cleared
-     * @param {boolean} [clear] If `true` the history is cleared
+     * @param {boolean?} clear If `true` the history is cleared
      */
-    store.reset = (clear) => {
+    const reset = (clear?: boolean) => {
         inAction = true
         index = 0
         store.set(states[index])
@@ -117,85 +151,63 @@ export const undoable = (initial, capacity, accept) => {
      * Get the number of saved states
      * @return {number}
      */
-    store.length = () => {
+    const length = () => {
         return states.length
     }
 
-    return store
-}
-
-/**
- * Check is an object have a function
- *
- * @internal
- * @private
- * @param {Object|*} anObject The object to test
- * @param {string} name The name of the function to test
- * @return {boolean}
- */
-const functionSupported = (anObject, name) => {
-    if (typeof anObject !== "object") {
-        return false
+    return {
+        subscribe: store.subscribe,
+        set: store.set,
+        update: store.update,
+        undo,
+        redo,
+        canUndo,
+        canRedo,
+        reset,
+        length,
     }
-    if (!Object.keys(anObject).includes(name)) {
-        return false
-    }
-    return typeof anObject[name] === "function"
-
 }
 
 /**
  * Change the value to the previous saved state
- * @param {Readable<*>} undoableStore The store to use
+ * @param {UndoableStore<*>} undoableStore The store to use
  */
-export const undo = (undoableStore) => {
-    if (functionSupported(undoableStore, "undo")) {
-        undoableStore.undo()
-    }
+export function undo<T> (undoableStore: UndoableStore<T>): void {
+    undoableStore.undo()
 }
 
 /**
  * Change the value to the next saved state
- * @param {Readable<*>} undoableStore The store to use
+ * @param {UndoableStore<*>} undoableStore The store to use
  */
-export const redo = (undoableStore) => {
-    if (functionSupported(undoableStore, "redo")) {
-        undoableStore.redo()
-    }
+export function redo<T> (undoableStore: UndoableStore<T>): void {
+    undoableStore.redo()
 }
 
 /**
  * Indicate if the store value can be revert to a previous state
- * @param {Readable<*>} undoableStore The store to use
+ * @param {UndoableStore<*>} undoableStore The store to use
  * @return {boolean}
  */
-export const canUndo = (undoableStore) => {
-    if (functionSupported(undoableStore, "canUndo")) {
-        return undoableStore.canUndo()
-    }
-    return false
+export function canUndo<T> (undoableStore: UndoableStore<T>): boolean {
+    return undoableStore.canUndo()
 }
 
 /**
  * Indicate if the store value can be change to a next state
- * @param {Readable<*>} undoableStore The store to use
+ * @param {UndoableStore<*>} undoableStore The store to use
  * @return {boolean}
  */
-export const canRedo = (undoableStore) => {
-    if (functionSupported(undoableStore, "canRedo")) {
-        return undoableStore.canRedo()
-    }
-    return false
+export function canRedo<T> (undoableStore: UndoableStore<T>): boolean {
+    return undoableStore.canRedo()
 }
 
 /**
  * Revert the value of the store to the oldest state.
  * If the second parameter is `true`, then the store state history is cleared
- * @param {Readable<*>} undoableStore The store to use
- * @param {boolean} [clear] If `true` the history is cleared
+ * @param {UndoableStore<*>} undoableStore The store to use
+ * @param {boolean} clear If `true` the history is cleared
  */
-export const reset = (undoableStore, clear) => {
-    if (functionSupported(undoableStore, "reset")) {
-        undoableStore.reset(clear)
-    }
+export function reset<T> (undoableStore: UndoableStore<T>, clear?: boolean): void {
+    undoableStore.reset(clear)
 }