+++ CHANGELOG.md
@@ -6,9 +6,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.1.0]
+
+### Added
+
+- Functional tests
+
+### Changed
+
+- Migrate the code to TypeScript
+
 ## [1.0.0] - 2020-09-12
 
 First version
 
-[Unreleased]: https://github.com/MacFJA/svelte-undoable/compare/1.0.0...HEAD
+[Unreleased]: https://github.com/MacFJA/svelte-undoable/compare/1.1.0...HEAD
+[1.1.0]: https://github.com/MacFJA/svelte-undoable/releases/tag/1.0.0...1.1.0
 [1.0.0]: https://github.com/MacFJA/svelte-undoable/releases/tag/1.0.0